requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
crispy-bootstrap5==2024.10
diff-match-patch==20241021
Django==4.2
django-admin-interface==0.29.1
django-appconf==1.0.6
django-colorfield==0.11.0
django-compressor==4.5.1
django-crispy-forms==2.3
django-extensions==3.2.3
django-filter==25.1
django-import-export==4.3.8
django-js-asset==3.1.2
django-mptt==0.17.0
django-multiselectfield==0.1.13
django-registration-redux==2.13
django-simple-history==3.10.1
django-tables2==2.7.0
django-tinymce==4.1.0
django-user-sessions==2.0.0
easy-thumbnails==2.10
idna==3.10
pdfkit==1.0.0
pillow==11.3.0
psycopg2-binary==2.9.10
python-decouple==3.8
python-slugify==8.0.4
rcssmin==1.1.2
reportlab==4.4.2
requests==2.32.4
rjsmin==1.2.2
setuptools==80.9.0
sqlparse==0.5.3
tablib==3.8.0
text-unidecode==1.3
urllib3==2.5.0
